Magento Plugin sends an invalid return_url
This url should NOT be the same as decline as it causes the order to become declined even after it has already been approved. This happens when the user clicks the back button in the receipt screen and is then redirected to the return_url which is the same as the decline_url.